Creation of repo for independent code development for various projects
Repository created to be able to develop and test functions for various projects within a single testbed project, independently of the repositories corresponding to these projects, preventing such development from interfering with development of the project repo such that proper behavior is confirmed before the code is added to it.

This is particularlu useful for major modifications of a project's functions or structure.  Modifications that will leave a program non-functional are not desirable inside the master repo, especially if there is other concurrent development of this repo, so rather than doing this within a separate branch of the repo, it is done within a single development testbed program which is designed for such development and can be done in Visual Studio.  Additional project repo development is added to this program by adding a .h/.cpp for this program and adding these to the list of included items in the master source file (containing main()).  This has been especially useful for the pCT project's redesigned console printing routines and the development of the AKS prime number algorithm.