Encapsulated code in main() in single test function for each project optionally called via bool control
Created a single test function for each project such that these are the only functions called in main() now, each with a single bool argument that controls whether it is called.  These project testing functions are then further abstracted into individual functions for tests of separate aspects of the project, which themselves are also optionally called via a single bool parameter.

The project testing functions and the project aspect testing functions they contain are then optionally called based on the bool/#define parameters defined in the top level testbed file containing main(), allowing this control to be performed from a single location.