Modified project dependency hierarchy w/ preprocessing & global/common vars/prototypes further separated into separate files
Partitioned content of code_development.h into additional separate headers directives.h, externs.h, and globals.h.  Preprocessing #includes/#defines are placed in directives.h, externally defined variables and functions are placed within externs.h, and globals.h is currently empty.  The goal is to move the global variable definitions to globals.h, but an appropriate implementation has not been found thus far, so these remain within common.h/.cpp.  The first globals.h attempt moved various definitions from common.h/.cpp into it, but a corresponding globals.cpp was not created, which is likely the reason the behavior with common.h/.cpp could not be duplicated with globals.h alone.  If so, globals.h/.cpp will likely be abandoned/removed and the desired behavior implemented in common.h/.cpp.  The variables/function prototypes of a particular testing project are placed in the corresponding header file such that it is straightforward and easy to copy the forward declarations and function implementations from the testing project to the actual project it is developed for.  Hence, difficulties arise with respect to enum and typedef definitions since the extern statements, placed in the globally accessible files making it possible to leave the variable/function definitions in the pCT testing project files, cannot be used for enums/typedefs.  It may not be possible to preserve the independence of the pCT testing project and make some of its variables/functions (e.g. the colored text print statements and program exit prompt/query functions) available for use with any of the other separate project testing routines, but it may not be possible without moving some enums/typedefs to the common.h/.cpp files and and removing/modifying their implementation in the pCT_... files.  Future commits will seek to investigate this topic.

The code in this commit has purposefully not been cleaned, not just from this hierarchal development, but also from the previous development/debugging of the pCT testing's slice_mergining_testing() which developed errors due to inappropriate template function parameterization and std::string/char* manipulation and argument passing/processing.  Additional development code remains commented out within several of the common.tem template functions (e.g. newly added string_cat()).  Subsequent commits will remove the commented lines of code used during development/debugging now that these appear to be functioning properly, but leaving the ability to easily retrieve and repeat the debugging code statements preserves this code for future debugging of these and similar functions should the need arise.