Added global template/variadic function allowing program exit after specified # of calls of function passed to it
Added a template/variadic function 'control_lkill' which accepts a function name, an integer used to set a counter variable used to control an early exit of a program, and a variable argument list 'Args... args_list'.  The function reference argument is defined to accept the argument list 'args_list' passed as the trailing arguments of 'control_lkill' as its argument, such that when calling 'control_lkill' with this function reference and the program exit control integer, the arguments for the referenced function should be appended to 'control_lkill''s list of arguments after the program exit control integer.  This program exit control integer is then assigned to the global variables __KILL_LOOP_INDEX and __KILL_FCALL_INDEX prior to the call of the referenced function, then after this function returns, these 2 global variables are reassigned the value INT_MAX such that the specified program exit control integer does not impact subsequent execution in cases where its value did not initiate program termination within that function.

The function uses this integer to initiate program termination in various ways, typically by adding statements to set/increment some counter variable and a program exit call conditioned on a comparison between this counter and the program exit control integer.  A common example would be a for loop index counter such that the program terminates after the specified # of loops.  The current usage (applied to control the # of merged image files) is to control the # of times a function called within a for loop reaches its innermost conditional branch, which may not happen on every call, if ever.  In this case, the global variable __fcall_index is reset in the referenced function but it's incremented within the helper function being called within the referenced function's for loop.  Hence, the usage of the control integer need not be within the function passed to 'control_lkill', numerous functions may contain statements dependent on the control integer with complicated relationships/interactions, preventing unintended interactions is the reason for the 2 global index threshold variables being assigned the value of the control integer immediately before calling the referenced function and resetting them immediately after it returns.

Therefore, such counter control and conditional program exit statements can be added to all functions where early program termination is desired.  This is particularly helpful for developing/debugging code containing large index for loops and/or code with potentially dangerous side effects, both of which apply in the slice merging procedure this procedure was developed for.  Slice merging is applied to every folder found within the specified directory, each corresponding to an individual reconstruction, which is currently a large list.  In addition, the file reading/writing performed can have potentially damaging effects.  During development, the process for checking if the merged file already existed resulted in those that did being overwritten with an empty file, damage that was corrected but given the need to interact with the component slice files which cannot so easily be recovered, it is important to constrain any potentially negative side effects to as few reconstructions as possible until proper behavior has been confirmed.  The initial implementation simply created a duplicate of the relevant function with an additional integer argument for the control integer, but given the usefulness of the procedure and the flexibility of template functions with variadic argument type lists, it was reimplemented as a single function that can be applied to any function.

In the previous commit, the project dependency hierarchy was modified to separate preprocessor directives and forward declarations of global variables/functions etc., but the corresponding statements in the files where these were previously located were simply commented out.  This commit removed these and performed additional code cleaning as well.